feat(mission): :necktie: Moved dynamic course content into a separate component
